backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
numpy
scikit-learn
scipy
statsmodels
python-multipart
openai
python-dotenv

# Optional — only needed for DATA_SOURCE=db
# sqlalchemy
# psycopg2-binary          # PostgreSQL
# pyodbc                   # MSSQL (also needs ODBC driver installed)
# cx_oracle                # Oracle

# Optional — only needed for DATA_SOURCE=api
# httpx
